Update main.cpp
Almost perfect, after debugging, and adding a few features like default arguments; this is for what position to look at in the string; added error checking, and a sentence that had numbers, punctuation, spaces, and special characters in use. 

if needed change the default argument of "spaces" to change the number of what char will be looked at.